cleaning up prefabs and code
