+++ Assets/_Scripts/Level/Brains/NPCController.cs
@@ -1,5 +1,4 @@
 
-using System;
 using System.Collections.Generic;
 using UnityEngine;
